requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiogram-broadcaster==0.0.7
aiohttp==3.8.5
aiolimiter==1.0.0
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.5.0
antiorm==1.2.1
anyio==3.6.2
APScheduler==3.9.1.post1
ascii-magic==1.6
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
behave==1.2.6
cachetools==5.2.1
cachier==1.5.3
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
comm==0.1.2
config==0.5.1
contourpy==1.0.6
coverage==7.0.5
cryptography==39.0.0
cycler==0.11.0
DateTime==5.1
db==0.1.1
db-sqlite3==0.0.1
debugpy==1.6.5
decorator==4.4.2
Django==4.2
entrypoints==0.4
executing==1.2.0
Faker==19.13.0
ffmpeg-python==0.2.0
Flask==2.2.2
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
httpcore==0.16.3
httpx==0.24.1
huepy==1.2.1
idna==3.4
imageio==2.24.0
imageio-ffmpeg==0.4.8
iniconfig==2.0.0
instabot==0.117.0
instaloader==4.9.5
ipykernel==6.19.4
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==7.4.8
jupyter_core==5.1.2
kiwisolver==1.4.4
loguru==0.5.3
magic-filter==1.0.11
MarkupSafe==2.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
mock==5.0.1
MouseInfo==0.1.3
moviepy==1.0.3
multidict==6.0.4
mysql-connector-python==8.2.0
nest-asyncio==1.5.6
numpy==1.24.1
packaging==21.3
pandas==1.5.2
parse==1.19.0
parse-type==0.6.0
parso==0.8.3
pathtools==0.1.2
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
plumbum==1.7.1
plyer==2.1.0
portalocker==2.6.0
proglog==0.1.10
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyGetWindow==0.0.9
Pygments==2.14.0
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
pymongo==3.12.1
PyMsgBox==1.0.9
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyPDF2==3.0.1
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.28
pyshorteners==1.0.1
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-env==1.0.0
python-rex==0.4
python-telegram-bot==20.5
pytube==15.0.0
pytube3==9.6.4
pytweening==1.0.4
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.4
pywin32==305
pywin32-ctypes==0.2.2
PyYAML==5.4.1
pyzmq==24.0.1
requests==2.28.2
requests-toolbelt==0.10.1
responses==0.22.0
rfc3986==1.5.0
schedule==1.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.3.2.post1
speedtest==0.0.1
speedtest-cli==2.1.3
sqlparse==0.4.3
stack-data==0.6.2
telebot==0.0.4
TelegramBotAPI==0.3.2
telepot==12.7
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
types-toml==0.10.8.1
typing_extensions==4.7.1
tzdata==2022.7
tzlocal==4.2
updater==0.3.2
urllib3==1.26.13
watchdog==2.2.1
wcwidth==0.2.5
Werkzeug==2.2.2
wikipedia==1.4.0
win32-setctime==1.1.0
yarl==1.8.2
youtube-dl==2021.12.17
zope.interface==6.0
